apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.12
aiofiles==24.1.0
aiosmtplib==3.0.2
httpx==0.27.2
python-dotenv==1.0.1
pyotp==2.9.0
websockets==13.1
